EspalexaWithWebServer example: Don't use bundled WebServer library
Use the copy of the WebServer library that's already in the Espalexa library, rather than a separate copy located in the sketch folder. Use of one copy of the WebServer library by the sketch and another by the Espalexa library was causing multiple definition errors.

It's necessary to move the include of Espalexa.h before the WebServer.h include so that the Espalexa library folder is added to the include search path.